@@ -9,6 +9,7 @@ class ExportTarget
 {
     private $fs;
     private $file = '';
+    private $url = '';
     private $content = '';
 
     public function __construct(FS $fs = null) {
@@ -23,6 +24,14 @@ class ExportTarget
         return $this->file;
     }
 
+    public function setExportUrl(string $url): void {
+        $this->url = $url;
+    }
+
+    public function getExportUrl(): string {
+        return $this->url;
+    }
+
     public function setContent(string $content): void {
         $this->content = $content;
     }
@@ -34,7 +43,7 @@ class ExportTarget
     public function save(): void {
         $result = $this->fs->write_file($this->getExportFile(), $this->getAsText());
         if (!$result) {
-            throw new FileWriteFailed();
+            throw new FileWriteFailed(spf_('Failed to write %s', $this->getExportFile()));
         }
     }
 }
