@@ -2,11 +2,9 @@
 
 namespace LnBlog\Export;
 
-use Blog;
-
 interface Exporter
 {
     public function setExportOptions(array $options): void;
     public function getExportOptions(): array;
-    public function export(Blog $blog, ExportTarget $target): void;
+    public function export($entity, ExportTarget $target): void;
 }
