@@ -10,7 +10,7 @@ use LnBlog\Export\Translators\WordPress\ReplyTranslator;
 use LnBlog\Storage\UserRepository;
 use SimpleXMLElement;
 
-class WordPressExporter implements Exporter
+class WordPressExporter extends BaseExporter implements Exporter
 {
     const EMPTY_EXPORT_XML = '<?xml version="1.0" encoding="UTF-8" ?>
 <rss version="2.0"
@@ -30,28 +30,9 @@ class WordPressExporter implements Exporter
         'wp' => "http://wordpress.org/export/1.2/",
     ];
 
-    private $globals;
-    private $user_repo;
-
-    private $options = [];
-
-    public function __construct(
-        GlobalFunctions $globals = null,
-        UserRepository $user_repo = null
-    ) {
-        $this->globals = $globals ?: new GlobalFunctions();
-        $this->user_repo = $user_repo ?: new UserRepository();
-    }
-
-    public function setExportOptions(array $options): void {
-        $this->options = $options;
-    }
-
-    public function getExportOptions(): array {
-        return $this->options;
-    }
-
-    public function export(Blog $blog, ExportTarget $target): void {
+    public function export($entity, ExportTarget $target): void {
+        /** @var Blog $blog */
+        $blog = $entity;
         $xml = new SimpleXMLElement(self::EMPTY_EXPORT_XML);
         $channel = $xml->addChild('channel');
         $blog_translator = new BlogTranslator($this->globals, $this->user_repo);
@@ -61,7 +42,7 @@ class WordPressExporter implements Exporter
         $this->translateArticles($channel, $blog->getArticles());
         $this->translateDrafts($channel, $blog->getDrafts());
 
-        $content = $this->prettyPrintXml($xml->asXML());
+        $content = self::prettyPrintXml($xml->asXML());
         $target->setContent($content);
     }
 
@@ -102,14 +83,5 @@ class WordPressExporter implements Exporter
         foreach ($replies as $reply) {
             $reply_translator->translate($item, $reply, ['comment_id' => $id++]);
         }
-    }
-
-    private function prettyPrintXml(string $xml): string {
-        $dom = new \DOMDocument('1.0');
-        $dom->preserveWhiteSpace = true;
-        $dom->formatOutput = true;
-        $dom->loadXML($xml);
-        $pretty_xml = $dom->saveXML();
-        return $pretty_xml;
     }
 }
