@@ -42,12 +42,18 @@ class BlogExportForm extends BaseForm
                 $this->blogidExists(),
                 $this->blogidToBlog()
             ),
+            'format' => new FormField(
+                'format',
+                new SelectRenderer(ExporterFactory::SUPPORTED_FORMATS),
+                $this->formatValid()
+            ),
         ];
     }
 
     protected function doAction(): ExportTarget {
         $default_file_name = 'export-' . date('Y-m-d_H_i_s') . '.xml';
-        $exporter = $this->exporter_factory->create(ExporterFactory::EXPORT_WORDPRESS);
+        $format = $this->fields['format']->getValue();
+        $exporter = $this->exporter_factory->create($format);
         $blog = $this->fields['blog_id']->getValue();
         $export_target = new ExportTarget();
         $export_target->setExportFile($default_file_name);
@@ -66,6 +72,15 @@ class BlogExportForm extends BaseForm
     private function blogidToBlog(): callable {
         return function (string $value): Blog {
             return $this->repository->get($value);
+        };
+    }
+
+    private function formatValid(): callable {
+        return function (string $value) {
+            if (!isset(ExporterFactory::SUPPORTED_FORMATS[$value])) {
+                return [spf_('Invalid export format "%s"', $value)];
+            }
+            return [];
         };
     }
 }
